Added support for additional data from Smart Hub 2 devices.
Namely :

'ConnectionType', 'ParentPhysAddress', 'ParentName'

Connection Type is Ether|5G|2.4G (G->Ghz)

ParentName is the name of the device the client is connecting through, i.e. the name of the disk or the core router itself.
ParentPhyAddress is the Mac Address of the  Parent.

Disks, the core router and inactive clients will have these additional fields set to 'Unknown'.